Removed the login functionality and db models to auth.py and models.py
