+++ main.py
@@ -11,86 +11,37 @@ from wtforms import StringField, PasswordField, SubmitField
 from wtforms.validators import InputRequired, Length, ValidationError
 from flask_bcrypt import Bcrypt
 
+from auth import auth as auth_blueprint
+from models import db, User
+
 
 app = Flask(__name__)
+bcrypt = Bcrypt(app)
 basedir = os.path.abspath(os.path.dirname(__file__))
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///'+os.path.join(basedir, 'database.db')
-
-db = SQLAlchemy(app)
-bcrypt = Bcrypt(app)
 app.config['SECRET_KEY'] = 'thisissecretkey'
 SESSION_TYPE = "redis"
 PERMANENT_SESSION_LIFETIME = 1800
 
+db.init_app(app)
+
 login_manager = LoginManager()
 login_manager.init_app(app)
-login_manager.login_view = 'login'
+login_manager.login_view = 'auth.login'
+
+app.register_blueprint(auth_blueprint)
 
 @login_manager.user_loader
 def load_user(user_id):
     return User.query.get(int(user_id))
 
-app.config.update(SECRET_KEY=os.urandom(24))
+#app.config.update(SECRET_KEY=os.urandom(24))
 
-final_list = []
+
 
 # Create the User table if it doesn't exist
 with app.app_context():
     db.create_all()
-
-@login_manager.user_loader
-def load_user(user_id):
-    return User.query.get(int(user_id))
-
-#user DB
-class User(db.Model, UserMixin):
-    id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String(20), nullable=False, unique=True)
-    password = db.Column(db.String(80), nullable=False)
-
-class RegisterForm(FlaskForm):
-    username = StringField(validators=[
-                           InputRequired(), Length(min=4, max=20)], render_kw={"placeholder": "Přihlašovací jméno"})
-
-    password = PasswordField(validators=[
-                             InputRequired(), Length(min=8, max=20)], render_kw={"placeholder": "Heslo"})
-
-    submit = SubmitField('Register')
-
-    def validate_username(self, username):
-        existing_user_username = User.query.filter_by(
-            username=username.data).first()
-        if existing_user_username:
-            raise ValidationError(
-                'That username already exists. Please choose a different one.')
-
-class LoginForm(FlaskForm):
-    username = StringField(validators=[
-                           InputRequired(), Length(min=4, max=20)], render_kw={"placeholder": "Přihlašovací jméno"})
-
-    password = PasswordField(validators=[
-                             InputRequired(), Length(min=8, max=20)], render_kw={"placeholder": "Heslo"})
-
-    submit = SubmitField('Přihlásit')
-
-
-@app.route('/login', methods=['GET', 'POST'])
-def login():
-    form = LoginForm()
-    if form.validate_on_submit():
-        user = User.query.filter_by(username=form.username.data).first()
-        if user:
-            if bcrypt.check_password_hash(user.password, form.password.data):
-                login_user(user)
-                return redirect(url_for('index'))
-    return render_template('login.html', form=form)
-
-
-@app.route('/logout', methods=['GET', 'POST'])
-@login_required
-def logout():
-    logout_user()
-    return redirect(url_for('login'))
 
 
 @app.route("/", methods=['GET', 'POST'])
@@ -136,6 +87,10 @@ def shortterm_memory_reading():
     print(type(combinations))
 
     return render_template('pages/shortterm-memory-reading.html', combinations=combinations)
+
+
+final_list = []
+
 
 @app.route("/shortterm-memory-listening")
 @login_required
@@ -394,10 +349,12 @@ def iq():
 
 
 @app.route("/circles")
+@login_required
 def circles():
     return render_template("pages/circles.html")
 
 @app.route("/rectangles")
+@login_required
 def rectangles():
     return render_template("pages/rectangles.html")
