+++ templates/base.html
@@ -61,7 +61,7 @@
           </li>
 
           <li class="nav-item">
-            <a class="nav-link" href="{{ url_for('logout') }}" style="color: red;">Odhlásit</a>
+            <a class="nav-link" href="{{ url_for('auth.logout') }}" style="color: red;">Odhlásit</a>
           </li>
 
         </ul>
